PCA plot

PCA plot

Usage

create_PCA_plot(
  dat,
  type = "sample_type",
  types_to_display = "all",
  threshold = NULL,
  interactive = TRUE
)

Arguments

dat

A raw_data object, the output of the read_data() function. The data should be completed and filtered to include only samples of type "Sample".

type

A character denoting which type of PCA plot should be created. Default is "sample_type", which makes a plot for quality control. Type "group" creates a PCA plot with respect to the groups of samples with type 'Sample'. Type "biplot" adds eigenvectors to the PCA for quality control.

types_to_display

A vector of sample type names specifying which types should be shown on the plot when type = "sample_type". Defaults to all.

threshold

A value indicating the minimum correlation between a variable and any component, required for this variable to be included on the PCA biplot.

interactive

logical indicating whether the plot should be interactive. If TRUE, the plot includes interactive tooltips.

Examples

path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
#> Completing 109 < LOD values...
#> Completing 6 < LLOQ values...
#> Completing 9 < ULOQ values...
create_PCA_plot(dat)
#> Joining with `by = join_by(tmp_id)`
#> Too few points to calculate an ellipse
#> Too few points to calculate an ellipse
#> Too few points to calculate an ellipse
#> Too few points to calculate an ellipse
#> Warning: Removed 4 rows containing missing values or values outside the scale range
#> (`geom_path()`).

create_PCA_plot(dat, type = "biplot", threshold = 0.3)
#> Joining with `by = join_by(tmp_id)`

dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
#> Completing 109 < LOD values...
#> Completing 6 < LLOQ values...
#> Completing 9 < ULOQ values...
create_PCA_plot(dat, type = "group")
#> Joining with `by = join_by(tmp_id)`