Boxplots of individual metabolite values before and after imputation

create_boxplot() returns boxplots based on the values of the specified metabolite before and after imputation.

Usage

create_boxplot(dat, metabolite, interactive = TRUE)

Arguments

dat

a raw_data object, the output of the read_data() function. The data have to be completed, for example using the complete_data() function.

metabolite

a name of the metabolite of interest.

interactive

logical. If TRUE (default), a ggiraph interactive plot is returned; otherwise, a standard ggplot object is produced.

Examples

path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
#> Completing 109 < LOD values...
#> Completing 6 < LLOQ values...
#> Completing 9 < ULOQ values...
create_boxplot(dat, "C0")
#> Warning: Removed 2 rows containing non-finite outside the scale range
#> (`stat_boxplot()`).