Qqplots of individual metabolite values before and after imputation

Qqplots of individual metabolite values before and after imputation

Usage

create_qqplot(dat, metabolite, interactive = TRUE)

Arguments

dat

A raw_data object, the output of the read_data() function. The data should be completed and filtered to include only samples of type "Sample".

metabolite

A name of metabolite of interest.

interactive

logical indicating whether the plot should be interactive. If TRUE, the plot includes interactive tooltips.

Examples

path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
#> Completing 109 < LOD values...
#> Completing 6 < LLOQ values...
#> Completing 9 < ULOQ values...
create_qqplot(dat, "C0")
#> Warning: Removed 2 rows containing non-finite outside the scale range (`stat_qq()`).
#> Warning: Removed 2 rows containing non-finite outside the scale range
#> (`stat_qq_line()`).