PCA plot

create_PCA_plot() returns the scatterplot or biplot, based on the PCA.

Usage

create_PCA_plot(
  dat,
  type = "scatterplot",
  group_by = "sample_type",
  types_to_display = "all",
  threshold = 0,
  interactive = TRUE
)

Arguments

dat

a raw_data object, the output of the read_data() function. The data have to be completed, for example using the complete_data() function.

type

a character denoting which type of PCA plot should be created. Default to "scatterplot". Type "biplot" shows eigenvectors.

group_by

character; default to "sample_type", which makes a plot for quality control. When set as "group", a PCA plot with respect to the groups of samples with type 'Sample' is created.

types_to_display

a vector of sample type names specifying which types should be shown on the plot when type = "sample_type". Defaults to all.

threshold

a value indicating the absolute minimum correlation between a metabolite and any of the first two components, required for this metabolite to be included on the biplot.

interactive

logical. If TRUE (default), a ggiraph interactive plot is returned; otherwise, a standard ggplot object is produced.

Examples

path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
#> Completing 109 < LOD values...
#> Completing 6 < LLOQ values...
#> Completing 9 < ULOQ values...
create_PCA_plot(dat)
#> Too few points to calculate an ellipse
#> Too few points to calculate an ellipse
#> Too few points to calculate an ellipse
#> Too few points to calculate an ellipse

create_PCA_plot(dat, type = "biplot", threshold = 0.3)

dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
#> Completing 109 < LOD values...
#> Completing 6 < LLOQ values...
#> Completing 9 < ULOQ values...
create_PCA_plot(dat, group_by = "group")

create_PCA_plot(dat, type = "biplot", group_by = "group", threshold = 0.3)